Commit Number: 12
Commit Message: Begin work on command system for user-app interaction

Commit Description: Started work on the command system to enable users
to interact with the app via commands. Created necessary classes,
designed the system architecture, and implemented some of the methods.
This is just the beginning of the development process, and I'm looking
forward to continuing this work.